ntp: don't remove original ntp.conf if dhcpcd is not included in DISTRO_FEATURES

+++ recipes-support/ntp/ntp_%.bbappend
@@ -2,8 +2,10 @@ inherit relative_symlinks
 
 PACKAGECONFIG_remove_karo-minimal = "openssl"
 
-FILES_${PN}_remove = "${sysconfdir}/ntp.conf"
+FILES_${PN}_remove = "${@ bb.utils.contains('DISTRO_FEATURES', 'dhcpcd', "${sysconfdir}/ntp.conf", "", d)}"
 
 do_install_append() {
-    rm -f ${D}${sysconfdir}/ntp.conf
+    if ${@ bb.utils.contains('DISTRO_FEATURES', 'dhcpcd', "true", "false", d)};then
+        rm -f ${D}${sysconfdir}/ntp.conf
+    fi
 }
